refactored to send MessageEvent, cancellation logic

+++ src/main/java/com/electronics/store/order_service/persistence/service/OrderService.java
@@ -2,11 +2,20 @@ package com.electronics.store.order_service.persistence.service;
 
 import com.electronics.store.order_service.controllers.dto.RequestItem;
 import com.electronics.store.order_service.controllers.misc.CreateOrderRequest;
+import com.electronics.store.order_service.events.OrderApplicationEvent;
+import com.electronics.store.order_service.events.OrderCancelledEvent;
 import com.electronics.store.order_service.events.OrderCreatedEvent;
 import com.electronics.store.order_service.grpc.Item;
+import com.electronics.store.order_service.grpc.ItemService;
 import com.electronics.store.order_service.persistence.model.Order;
+import com.electronics.store.order_service.persistence.model.OrderEvent;
 import com.electronics.store.order_service.persistence.model.OrderItem;
 import com.electronics.store.order_service.persistence.repositories.OrderRepository;
+import com.electronics.store.order_service.persistence.service.exceptions.NotEnoughItemsException;
+import com.electronics.store.order_service.persistence.service.exceptions.OrderCancellationIsNotAllowedException;
+import com.electronics.store.order_service.persistence.service.exceptions.OrderIsAlreadyCancelledException;
+import com.electronics.store.order_service.persistence.service.exceptions.OrderEventNotFoundException;
+import com.electronics.store.order_service.validation.ItemValidator;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,14 +23,13 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.util.*;
 
-import static com.electronics.store.order_service.persistence.enums.OrderEventStatus.NEW;
-import static com.electronics.store.order_service.persistence.enums.OrderEventType.ORDER_CREATED;
-import static com.electronics.store.order_service.persistence.enums.OrderStatus.PENDING;
-import static com.electronics.store.order_service.persistence.mapping.Mapper.mapToOrderEvent;
-import static java.time.OffsetDateTime.now;
+import static com.electronics.store.order_service.persistence.enums.PublishmentStatus.*;
+import static com.electronics.store.order_service.persistence.enums.OrderEventType.*;
+import static com.electronics.store.order_service.persistence.mapping.EntityCreator.createOrder;
+import static com.electronics.store.order_service.persistence.mapping.EntityCreator.createOrderEvent;
+import static java.util.stream.Collectors.toSet;
 
 @Slf4j
 @Service
@@ -32,60 +40,68 @@ public class OrderService {
     private final OrderRepository orderRepository;
     private final OrderEventService orderEventService;
     private final ApplicationEventPublisher eventPublisher;
+    private final ItemService itemService;
+    private final ItemValidator itemValidator;
 
     public Optional<Order> findByOrderId(UUID orderId) {
         return orderRepository.findById(orderId);
     }
 
     @Transactional
-    public Order persist(@Valid CreateOrderRequest request, Map<UUID, Item> itemsByIds) {
-        final Order order = new Order();
-        order.setCustomerId(request.customerId());
-        order.setStatus(PENDING);
-        order.setCreatedAt(now());
-        order.setCurrency(request.currency());
-        order.setTotalPrice(getTotalPrice(itemsByIds));
-
-        final Set<OrderItem> orderItems = new HashSet<>(request.orderItems().size());
-        for (RequestItem requestItem : request.orderItems()) {
-            final Item actualItem = itemsByIds.get(requestItem.itemId());
-            orderItems.add(new OrderItem(null, requestItem.itemId(), actualItem.description(),
-                    requestItem.quantity(), actualItem.price(), actualItem.imageData(),
-                    actualItem.itemUrl(), order));
+    public Order persist(@Valid CreateOrderRequest request) {
+        final Map<UUID, Item> itemsByIds = itemService.getItemsByIds(getItemIds(request));
+        if (!itemValidator.isValid(request, itemsByIds)) {
+            throw new NotEnoughItemsException("Not enough items in inventory!");
         }
-        order.setItems(orderItems);
-
-        final Order savedOrder = orderRepository.save(order);
-        orderEventService.persist(mapToOrderEvent(savedOrder, ORDER_CREATED, NEW));
-        publishAfterOrderIsCreated(savedOrder.getId());
-        return savedOrder;
-    }
-
-    public void publishAfterOrderIsCreated(UUID orderId) {
-        final OrderCreatedEvent orderCreatedEvent = new OrderCreatedEvent(orderId);
-        log.info("Sending application OrderCreatedEvent: {}", orderCreatedEvent);
-        eventPublisher.publishEvent(orderCreatedEvent);
+        final Order order = orderRepository.save(createOrder(request, itemsByIds));
+        orderEventService.persist(createOrderEvent(order, ORDER_CREATED, NEW, () -> itemsByIds));
+        publishOrderEvent(new OrderCreatedEvent(order.getId()));
+        log.info("Order stored: {}", order);
+        return order;
     }
 
     @Transactional
-    public int deleteByOrderId(UUID orderId) {
-        int orderRows = orderRepository.removeById(orderId);
-        int orderEventRows = orderEventService.removeByOrderId(orderId);
-        if (orderRows != 1 && orderEventRows != 1) {
-            log.error("Expected to be removed only one Order and one OrderEvent but there were removed " +
-                    "orders: {}, orderEvents: {}", orderRows, orderEventRows);
+    public void cancelOrder(UUID orderId) {
+        final OrderEvent orderEvent = orderEventService.findLastByOrderId(orderId)
+                .orElseThrow(() -> new OrderEventNotFoundException("Order event with orderId: " + orderId + " not found!"));
+        log.info("Found the latest order event: {}", orderEvent);
+        if (NEW.equals(orderEvent.getStatus()) && ORDER_CREATED.equals(orderEvent.getEventType())) {
+            int orderRows = orderRepository.removeById(orderId);
+            int orderEventRows = orderEventService.removeByOrderId(orderId);
+            if (orderRows != 1 && orderEventRows != 1) {
+                log.error("Expected to be removed only one Order and one OrderEvent but there were removed " +
+                        "orders: {}, orderEvents: {}", orderRows, orderEventRows);
+            }
+            log.info("Order event with orderId: {} is cancelled", orderId);
+            return;
+        } else if (ORDER_CANCELLED.equals(orderEvent.getEventType())) {
+            log.info("Order event cancelling with orderId: {} has already been triggered", orderId);
+            throw new OrderIsAlreadyCancelledException("Order event with orderId: " + orderId + " is cancelled");
+        } else if (SHIPMENT_CREATED.equals(orderEvent.getEventType()) || SHIPMENT_COMPLETED.equals(orderEvent.getEventType())) {
+            log.info("Unable to cancel order event with orderId: {}, because shipment has already been started", orderId);
+            throw new OrderCancellationIsNotAllowedException("Unable to cancel order event with orderId: " + orderId +
+                    " because shipment has already been started");
         }
-        return orderRows;
+        final Order order = orderRepository.findById(orderId).orElseThrow(() -> new OrderEventNotFoundException("Order with orderId: " + orderId + " not found!"));
+        orderEventService.persist(createOrderEvent(order, ORDER_CANCELLED, NEW, () -> itemService.getItemsByIds(getItemIds(order))));
+        publishOrderEvent(new OrderCancelledEvent(orderId));
+        log.info("Order event with orderId: {} is being cancelled", orderId);
     }
 
     public List<Order> findByCustomerId(UUID customerId) {
         return orderRepository.findOrdersByCustomerId(customerId);
     }
 
-    private BigDecimal getTotalPrice(Map<UUID, Item> itemsByIds) {
-        return itemsByIds.values().stream()
-                .map(Item::price)
-                .reduce(BigDecimal::add)
-                .orElseThrow();
+    private void publishOrderEvent(OrderApplicationEvent applicationEvent) {
+        log.info("Sending application event: {}", applicationEvent);
+        eventPublisher.publishEvent(applicationEvent);
+    }
+
+    private Set<UUID> getItemIds(CreateOrderRequest request) {
+        return request.orderItems().stream().map(RequestItem::itemId).collect(toSet());
+    }
+
+    private Set<UUID> getItemIds(Order order) {
+        return order.getItems().stream().map(OrderItem::getItemId).collect(toSet());
     }
 }
